Finished refresh token implementation, more info in the Desc
The refresh token is fully implemented in the WelcomePage. I highlighted the part that needs to be copied over to other pages. Also use useCheckAuth() and SessionModal.
Had to edit the login/auth/status endpoint so it would return the state of the existence of the refresh token. Also added the refresh token state to other files where the state of the token is used. Made it so if the refresh token doesn't exist, it doesn't matter if the token does, it deletes both tokens and send you to the logout page.